+++ setup.cfg
@@ -25,7 +25,7 @@ packages = find:
 
 [options.entry_points]
 console_scripts =
-    tissuumaps-schema = tissuumaps_schema.__main__:main
+    tissuumaps-schema = tissuumaps_schema._cli:cli
 
 [flake8]
 max-line-length = 88
