+++ tissuumaps_schema/v01/__init__.py
@@ -1,4 +1,4 @@
-from ._v01 import VERSION
+from .base import VERSION
 from .project import Project
 
 __all__ = ["VERSION", "Project"]
